+++ backend/src/main/java/com/back/domain/proposal/proposal/dto/ProposalStateUpdateReqBody.java
@@ -1,6 +1,6 @@
-package com.back.domain.project.proposal.dto;
+package com.back.domain.proposal.proposal.dto;
 
-import com.back.domain.project.proposal.constant.ProposalStatus;
+import com.back.domain.proposal.proposal.constant.ProposalStatus;
 import jakarta.validation.constraints.NotNull;
 
 public record ProposalStateUpdateReqBody (
