+++ backend/src/main/java/com/back/domain/proposal/proposal/entity/Proposal.java
@@ -1,23 +1,18 @@
-package com.back.domain.project.proposal.entity;
+package com.back.domain.proposal.proposal.entity;
 
 import com.back.domain.freelancer.freelancer.entity.Freelancer;
 import com.back.domain.project.project.entity.Project;
-import com.back.domain.project.proposal.constant.ProposalStatus;
+import com.back.domain.proposal.proposal.constant.ProposalStatus;
 import com.back.global.jpa.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
 
 @Entity
 @Getter
