+++ backend/src/main/java/com/back/global/initdata/BaseInitData.java
@@ -13,7 +13,7 @@ import com.back.domain.member.member.entity.Member;
 import com.back.domain.member.member.service.MemberService;
 import com.back.domain.project.project.entity.Project;
 import com.back.domain.project.project.service.ProjectService;
-import com.back.domain.project.proposal.service.ProposalService;
+import com.back.domain.proposal.proposal.service.ProposalService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationRunner;
